Project dashboard (#85)
Re-orienting the service dashboard around projects. Playground users
will only have 1 project for the foreseeable future, but anchoring the
portal dashboard around projects makes it easier to offer project-level
operations, in particular, how to delete a project.

Users often deploy a test project to defang, open the portal, and then
wonder how to delete it. @lionello is working on giving portal
permission to delete projects in
https://github.com/DefangLabs/defang-mvp/pull/993. In the meantime,
we're adding some help text to the project card which shows how to
delete a project from the CLI.

This PR moves the following routes `/service` to `/projects` (and
introduces a redirect rule), but it keeps `/service/:service-name` in
place.

There are a few other style tweaks in this PR related to the sidebar